Scope host readiness helper
